pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-flux"
version = "0.1.0"
description = "CLI tool for running LLM batch processing jobs on Delta"
authors = [
    {name = "Rohan Marwaha", email = "rohan13@illinois.edu"}
]
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "tqdm>=4.66.1",
    "jsonschema>=4.20.0",
    "tenacity>=8.2.0",
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "pyyaml>=6.0.1"
]

[project.scripts]
aiflux = "aiflux.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"aiflux" = [
    "templates/models/*/*.yaml",
    "container/*.def",
    "slurm/*.sh"
] 